combined pass with csl-wrapper-to-csl
