+++ xdsl/tools/command_line_tool.py
@@ -101,11 +101,6 @@ def get_all_passes() -> dict[str, Callable[[], type[ModulePass]]]:
 
         return csl_stencil_bufferize.CslStencilBufferize
 
-    def get_csl_wrapper_lower_import():
-        from xdsl.transforms import csl_wrapper_lower_import
-
-        return csl_wrapper_lower_import.CslWrapperLowerImportPass
-
     def get_csl_stencil_to_csl_wrapper():
         from xdsl.transforms import csl_stencil_to_csl_wrapper
 
@@ -424,7 +419,6 @@ def get_all_passes() -> dict[str, Callable[[], type[ModulePass]]]:
         "convert-stencil-to-ll-mlir": get_convert_stencil_to_ll_mlir,
         "cse": get_cse,
         "csl-stencil-bufferize": get_csl_stencil_bufferize,
-        "csl-wrapper-lower-import": get_csl_wrapper_lower_import,
         "csl-stencil-to-csl-wrapper": get_csl_stencil_to_csl_wrapper,
         "csl-wrapper-to-csl": get_csl_wrapper_to_csl,
         "dce": get_dce,
